move notifiations to own component
